+++ src/generator/sphere.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 
 
-std::string trans(double, double, double);
+std::string trans(float, float, float);
 
 std::string sphere(float radius, int slices, int stacks) {
     if(radius <= 0.0f || slices <= 0 || stacks <= 0) {
@@ -15,16 +15,16 @@ std::string sphere(float radius, int slices, int stacks) {
     std::ostringstream os;
 
     int nPoints {0};
-    double deltaAlpha {2.0f * M_PI / slices};
-    double deltaBeta {M_PI / stacks};
+    float deltaAlpha { static_cast<float>(2.0f * M_PI / slices) };
+    float deltaBeta { static_cast<float>(M_PI / stacks) };
 
     for(int i {0}; i < stacks; i++) {
-        double beta {i * deltaBeta};
-        double nextBeta {beta + deltaBeta};
+        float beta {i * deltaBeta};
+        float nextBeta {beta + deltaBeta};
 
         for(int j {0}; j < slices; j++) {
-            double alpha {j * deltaAlpha};
-            double nextAlpha {alpha + deltaAlpha};
+            float alpha {j * deltaAlpha};
+            float nextAlpha {alpha + deltaAlpha};
 
             if(i < stacks-1) {
                 os << trans(beta, alpha, radius);
@@ -45,6 +45,6 @@ std::string sphere(float radius, int slices, int stacks) {
     return r.str();
 }
 
-std::string trans(double a, double b, double c) {
+std::string trans(float a, float b, float c) {
     return writePoint(c * sinf(a) * sinf(b), c * cosf(a), c * sinf(a) * cosf(b));
 }
